Changes to const and var dec in grammar, analyzer, core
