+++ .gitignore
@@ -135,3 +135,5 @@ dist
 .ds_store
 #py cache
 __pycache__/*
+
+old_src/
